+++ src/KSPTextureLoader/TextureHandle.cs
@@ -75,7 +75,6 @@ internal class TextureHandleImpl : IDisposable, ISetException, ICompleteHandler
     public Texture TakeTexture()
     {
         using var guard = this;
-
         var texture = GetTexture();
 
         if (RefCount == 1 && AssetBundle is null)
@@ -164,8 +163,16 @@ internal class TextureHandleImpl : IDisposable, ISetException, ICompleteHandler
         if (RefCount != 0)
             return;
 
-        // The destroy queue will clear out the textures at the end of the frame.
-        TextureLoader.Instance.QueueForDestroy(this);
+        if (!IsError && texture == null)
+        {
+            // Destroy the handle immediately if its texture has been taken.
+            Destroy();
+        }
+        else
+        {
+            // The destroy queue will clear out the textures at the end of the frame.
+            TextureLoader.Instance.QueueForDestroy(this);
+        }
     }
 
     internal uint Destroy(bool immediate = false)